requirements.txt
PySide6>=6.5.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0

# AI Semantic Search (CLIP)
torch>=2.0.0
transformers>=4.30.0

# Safe Delete (Recycle Bin)
Send2Trash>=1.8.0

# Fast Similarity Search (Faiss)
faiss-cpu>=1.7.0
